+++ p2p/peer_stats.go
@@ -26,7 +26,7 @@ type peerStat struct {
 // by dividing the amount by time, so the result score will represent how many bytes
 // were retrieved in 1 millisecond. This value will then be averaged relative to the
 // previous peerScore.
-func (p *peerStat) updateStats(amount uint64, duration time.Duration) {
+func (p *peerStat) updateStats(amount int, duration time.Duration) {
 	p.Lock()
 	defer p.Unlock()
 	averageSpeed := float32(amount)
